pingid_plotlydash/requirements.txt
dash
authlib
dash-auth
starlette==0.13.6
azure-keyvault-secrets 
azure-identity